requirements.txt
fastapi~=0.92.0
Jinja2~=3.1.2
uvicorn~=0.20.0
pymongo~=4.3.3
bcrypt~=4.0.1
PyJWT~=2.6.0
pydantic~=1.10.4
MyApplication~=0.1.0